Refactor src/components/AgendaItems/AgendaItemsContainer.module.css
