--- src/components/AgendaItems/AgendaItemsContainer.module.css
@@ -1,230 +1,0 @@
-.createModal {
-  margin-top: 20vh;
-  margin-left: 13vw;
-  max-width: 80vw;
-}
-
-.titlemodal {
-  color: var(--bs-gray-600);
-  font-weight: 600;
-  font-size: 20px;
-  margin-bottom: 20px;
-  padding-bottom: 5px;
-  border-bottom: 3px solid var(--bs-primary);
-  width: 65%;
-}
-
-.agendaItemsOptionsButton {
-  width: 24px;
-  height: 24px;
-}
-
-.agendaItemModal {
-  max-width: 80vw;
-  margin-top: 2vh;
-  margin-left: 13vw;
-}
-
-.iconContainer {
-  display: flex;
-  justify-content: flex-end;
-}
-.icon {
-  margin: 1px;
-}
-
-.errorIcon {
-  transform: scale(1.5);
-  color: var(--bs-danger);
-  margin-bottom: 1rem;
-}
-
-.greenregbtn {
-  margin: 1rem 0 0;
-  margin-top: 15px;
-  border: 1px solid var(--bs-gray-300);
-  box-shadow: 0 2px 2px var(--bs-gray-300);
-  padding: 10px 10px;
-  border-radius: 5px;
-  background-color: var(--bs-primary);
-  width: 100%;
-  font-size: 16px;
-  color: var(--bs-white);
-  outline: none;
-  font-weight: 600;
-  cursor: pointer;
-  transition:
-    transform 0.2s,
-    box-shadow 0.2s;
-  width: 100%;
-}
-
-.preview {
-  display: flex;
-  flex-direction: row;
-  font-weight: 900;
-  font-size: 16px;
-  color: rgb(80, 80, 80);
-}
-
-.view {
-  margin-left: 2%;
-  font-weight: 600;
-  font-size: 16px;
-  color: var(--bs-gray-600);
-}
-
-.previewFile {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  margin-top: 10px;
-}
-
-.previewFile img,
-.previewFile video {
-  width: 100%;
-  max-width: 400px;
-  height: auto;
-  margin-bottom: 10px;
-}
-
-.attachmentPreview {
-  position: relative;
-  width: 100%;
-}
-
-.closeButtonFile {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  background: transparent;
-  transform: scale(1.2);
-  cursor: pointer;
-  border: none;
-  color: #707070;
-  font-weight: 600;
-  font-size: 16px;
-  cursor: pointer;
-}
-
-.tableHeader {
-  background-color: var(--bs-primary);
-  color: var(--bs-white);
-  font-size: 16px;
-}
-
-.noOutline input {
-  outline: none;
-}
-
-.categoryContainer {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-  justify-content: center;
-}
-
-.categoryChip {
-  display: inline-flex;
-  align-items: center;
-  background-color: #e0e0e0;
-  border-radius: 16px;
-  padding: 0 12px;
-  font-size: 14px;
-  height: 32px;
-  margin: 5px;
-}
-
-.urlListItem {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 5px 0;
-}
-
-.urlIcon {
-  margin-right: 10px;
-}
-
-.deleteButton {
-  margin-left: auto;
-  padding: 2px 5px;
-}
-
-.urlListItem a {
-  text-decoration: none;
-  color: inherit;
-}
-
-.urlListItem a:hover {
-  text-decoration: underline;
-}
-
-.agendaItemRow {
-  border: 1px solid #dee2e6;
-  border-radius: 4px;
-  transition: box-shadow 0.2s ease;
-  background-color: #fff;
-}
-.agendaItemRow:hover {
-  background-color: #f0f0f0;
-}
-
-.dragging {
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.3);
-  z-index: 1000;
-  background-color: #f0f0f0;
-}
-
-.droppable {
-  background-color: #f9f9f9; /* Background color of droppable area */
-}
-
-.droppableDraggingOver {
-  background-color: #e6f7ff; /* Background color of droppable area while dragging over */
-}
-
-.tableHead {
-  background-color: #31bb6b !important;
-  color: white;
-  border-radius: 20px 20px 0px 0px !important;
-  padding: 20px;
-}
-
-@media (max-width: 768px) {
-  .createModal,
-  .agendaItemModal {
-    margin: 10vh auto;
-    max-width: 90%;
-  }
-
-  .titlemodal {
-    width: 90%;
-  }
-
-  .greenregbtn {
-    width: 90%;
-  }
-
-  /* Add more specific styles for smaller screens as needed */
-}
-
-@media (max-width: 576px) {
-  .createModal,
-  .agendaItemModal {
-    margin: 5vh auto;
-    max-width: 95%;
-  }
-
-  .titlemodal {
-    width: 100%;
-  }
-
-  .greenregbtn {
-    width: 100%;
-  }
-
-  /* Additional specific styles for even smaller screens */
-}
